A number guessing game in python
the game uses functions, conditional statements, and loops. It also uses
the random library to make the number random. For the easy mode, you get
infinite guesses. The hard mode only gives you three guesses. It tells
you whether you're close or not to the random number to help you guess
what it is.